Fixed Game.py Stalemate issue
Originally, if you run out of food and have no other ants besides your
queen, you could still keep playing. The issue with this is against a
random AI it can take a very long time to lose. Also, if your opponent
is in the same situation then there is no way for either of you to win
or lose.