Fixed bug where AI does not make a move
For some reason, in the case where the AI loses all of their attack
ants, instead of moving the queen and building a new ant, the AI simply
stops making moves entirely, not even an end turn move.

It is possible that somehow the moves that involve moving the queen or
building an ant are pruned away but then the AI should still pick a
move, however bad.